requirements.txt
python-telegram-bot==21.1
openai
psycopg2-binary